Move cutout functionality to docker env variables.
- Refactored main.py and utils.py so that the Epochs aren't all stored directly in browser.
