+++ robbie_viewer_server/utils.py
@@ -1,7 +1,12 @@
 from functools import lru_cache
+from os import stat
 from astropy.io import fits
 from astropy.io.votable import parse
 from astropy.wcs import WCS
+from astropy.coordinates import SkyCoord
+from astropy.nddata import Cutout2D
+
+import astropy.units as u
 
 import numpy as np
 import datetime
@@ -18,22 +23,26 @@ TOOLS = "box_select,lasso_select,help,pan,tap,wheel_zoom,reset"
 selected_circle = Circle(fill_alpha=1, fill_color="firebrick", line_color=None)
 nonselected_circle = Circle(fill_alpha=0.2, fill_color="blue", line_color=None)
 
-
 @lru_cache
 def load_mean_image(filename):
     hdu = fits.open(filename)#'results/mean_image.fits')
+    hdu[0].data = hdu[0].data
     return hdu
 
-
-def mean_image_data(hdu):
-    wcs = WCS(hdu[0].header)
+def mean_image_data(hdu, ra_ref=None, dec_ref=None, degrees_around_ref_coords=None):
+    wcs = WCS(hdu[0].header)   
     data = np.fliplr(hdu[0].data)
+    if ra_ref is not None:
+        cutout_position = SkyCoord(ra_ref*u.deg, dec_ref*u.deg, frame='icrs')
+        cutout = Cutout2D(hdu[0].data, cutout_position, degrees_around_ref_coords*u.deg, wcs=wcs, mode='partial')
+        wcs = cutout.wcs
+        data = np.fliplr(cutout.data)   
+        
     x,y = np.indices(data.shape[::-1])
     sky_map = wcs.pixel_to_world(x.reshape(-1),y.reshape(-1))
     ra = np.fliplr(sky_map.ra.degree.reshape(x.shape).T)
     dec = sky_map.dec.degree.reshape(y.shape).T
     return data, ra, dec
-
 
 def get_imdata(data, ra, dec, ei=None):
     if ra[0][0] > ra[0][-1]:
@@ -72,21 +81,32 @@ def get_imdata(data, ra, dec, ei=None):
         }
     return imdata
 
-
 def get_tabdata(fname):
     tab = parse(fname).get_first_table().to_table()
     df = tab.to_pandas()
     return df
 
 @lru_cache
-def get_joined_table_source(result_dir):
+def get_joined_table_source(result_dir, ra_ref=None, dec_ref=None, degrees_around_ref_coords=None):
     # most likely there is a solution that doesn't involve so many intermediaries!
     flux_table = get_tabdata(f"{result_dir}/flux_table.vot")
     stats_table = get_tabdata(f"{result_dir}/stats_table.vot")
-    joined = flux_table.join(stats_table.set_index('uuid'), on='uuid')
+
+    # Mask data based on provided RA and DEC ranges
+    if ra_ref is not None:
+        dec_min, dec_max = dec_ref - degrees_around_ref_coords/2, dec_ref + degrees_around_ref_coords/2
+        ra_min, ra_max = ra_ref - degrees_around_ref_coords/2, ra_ref + degrees_around_ref_coords/2
+        flux_table = flux_table.drop(flux_table[(flux_table['ref_ra'] > ra_max) + (flux_table['ref_ra'] < ra_min)].index)
+        flux_table = flux_table.drop(flux_table[(flux_table['ref_dec'] > dec_max) + (flux_table['ref_dec'] < dec_min)].index)
+    
+    # # Drop NaN
+    flux_table = flux_table.dropna()
+    stats_table = stats_table.dropna()
+
+    joined = flux_table.join(stats_table.set_index('uuid'), on='uuid')    
     df = joined
     source = ColumnDataSource(data=dict( (i,df[i]) for i in df.columns))
-
+    
     flux_cols = list(sorted([i for i in df.columns if i.startswith('peak_flux_')]))
     err_flux_cols = [f'err_{i}' for i in flux_cols]
     fluxes = df.set_index('uuid')[flux_cols]
@@ -105,14 +125,13 @@ def get_joined_table_source(result_dir):
         'epoch':epochs,
         'date':dates,
         'datetimes':datetimes,
-        'current':fluxes[df['uuid'][0]],
-        'current_upper':fluxes[df['uuid'][0]].values+err_fluxes[df['uuid'][0]].values,
-        'current_lower':fluxes[df['uuid'][0]].values-err_fluxes[df['uuid'][0]].values,
+        'current':fluxes[df['uuid'].iloc[0]],
+        'current_upper':fluxes[df['uuid'].iloc[0]].values+err_fluxes[df['uuid'].iloc[0]].values,
+        'current_lower':fluxes[df['uuid'].iloc[0]].values-err_fluxes[df['uuid'].iloc[0]].values,
     })
 
     lc_source = ColumnDataSource(data=lc)
     return source, lc_source
-
 
 def get_scatter_plots(source):
 
@@ -150,9 +169,12 @@ def get_scatter_plots(source):
     data_table = DataTable(source=source, columns=columns, scroll_to_selection=True, sortable=True, background='#111111')
     return left,right,data_table
 
-def get_mean_image_plot(source, result_dir):
+def get_mean_image_plot(source, result_dir, ra_ref=None, dec_ref=None, degrees_around_ref_coords=None):
     hdu = load_mean_image(f"{result_dir}/mean_image_reprojected.fits")
-    data, ra, dec = mean_image_data(hdu)
+    if ra_ref:
+        data, ra, dec = mean_image_data(hdu, ra_ref, dec_ref, degrees_around_ref_coords)
+    else:
+        data, ra, dec = mean_image_data(hdu)
     imdata = get_imdata(data,ra,dec)
 
     p = figure(
@@ -184,6 +206,42 @@ def get_mean_image_plot(source, result_dir):
     #p.grid.grid_line_width = 0.5
     return p
 
+def get_epoch_image_plots(source, epoch_source, num_epochs):
+    # Set up the figure
+    epochs = figure(
+        tools=TOOLS,
+        title="Epoch_0",
+        tooltips=[
+            ("value", "@image Jy/beam"),
+            ("RA", "@ra{0.00}°"),
+            ("DEC", "@dec{0.00}°")
+        ],
+        x_axis_label='RA',
+        y_axis_label='DEC',
+        x_range=(min(source.data['ref_ra']),  max(source.data['ref_ra'])),
+        y_range=(min(source.data['ref_dec']), max(source.data['ref_dec'])),
+    )
+
+
+
+    # Adjust hover tool to only work on image data
+    hover_tool = epochs.select(type=HoverTool)
+    hover_tool.names = ["epoch_image"]
+    epoch_slider = Slider(start=0, end=num_epochs-1, value=0, step=1, title="Epoch")
+
+    # must give a vector of image data for image parameter
+    epochs.image(source=epoch_source,
+            palette=palettes.mpl['Cividis'][256], name="epoch_image")
+
+
+    # Add the source circles
+    epochs.circle(source=source,
+                x='ref_ra', y='ref_dec',
+                radius=0.03, fill_color=None,
+                line_width=1.5, line_color='yellow')
+
+    return epochs, epoch_slider
+
 def get_light_curve_plot(source):
     tooltips = [("epoch","@date"),]
     lc_plot = figure(
@@ -213,63 +271,3 @@ def get_light_curve_plot(source):
         r.nonselection_glyph = nonselected_circle
 
     return lc_plot
-
-def get_epoch_image_plots(epoch_files, mean_source):
-    # Set up the figure
-    p = figure(
-        tools=TOOLS,
-        title="Epoch_0",
-        tooltips=[
-            ("value", "@image Jy/beam"),
-            ("RA", "@ra{0.00}°"),
-            ("DEC", "@dec{0.00}°")
-        ],
-        x_axis_label='RA',
-        y_axis_label='DEC',
-        x_range=(min(mean_source.data['ref_ra']),  max(mean_source.data['ref_ra'])),
-        y_range=(min(mean_source.data['ref_dec']), max(mean_source.data['ref_dec'])),
-    )
-
-    # Adjust hover tool to only work on image data
-    hover_tool = p.select(type=HoverTool)
-    hover_tool.names = ["epoch_image"]
-
-    # Make a data dictionary of each epoch with the _(int) format keys
-    data_dict = {}
-    for ei, epoch_file in enumerate(epoch_files):
-        hdu = load_mean_image(epoch_file)
-        data, ra, dec = mean_image_data(hdu)
-        imdata = get_imdata(data, ra, dec, ei=ei)
-        data_dict.update(imdata)
-    # Point to first image first
-    data_dict.update({
-            'image':data_dict["image_0"],
-            'ra':data_dict["ra_0"],
-            'dec':data_dict["dec_0"],
-            'x':data_dict["x_0"],
-            'y':data_dict["y_0"],
-            'dw':data_dict["dw_0"],
-            'dh':data_dict["dh_0"]})
-    source = ColumnDataSource(data=data_dict)
-
-    # must give a vector of image data for image parameter
-    p.image(source=source,
-            palette=palettes.mpl['Cividis'][256], name="epoch_image")
-    slider = Slider(start=0, end=len(epoch_files)-1, value=0, step=1, title="Epoch")
-    # Add the source circles
-    p.circle(source=mean_source,
-             x='ref_ra', y='ref_dec',
-             radius=0.03, fill_color=None,
-             line_width=1.5, line_color='yellow')
-
-    callback = CustomJS(args=dict(source=source, slider=slider, p=p),
-                        code="""
-        const data = source.data;
-        const i = slider.value;
-        p.title.text = 'Epoch_'+i.toString(10);
-        const image = data['image_'+i.toString(10)];
-        data['image'] = image;
-        source.change.emit();
-    """)
-    slider.js_on_change("value", callback)
-    return p, slider
